refactor: split the three oversized components into focused files
PersonalDataForm 281 to 158 lines: radio option data moved to OrderOptions, plus RecipientFields, RadioField, FieldError and a shared class module.
FixedHeader 246 to 76 lines: state and effects extracted into useHeaderState, markup split into BurgerButton, HeaderMenu, HeaderBuyButtons and HeaderIcons.
Product 205 to 51 lines: purchase logic moved into ProductPurchasePanel and wishlist/share into ProductActions, leaving Product as layout only.

Pure refactor, no markup or behaviour changes.
