+++ src/components/PageOrder/FieldError.tsx
@@ -1,0 +1,4 @@
+export function FieldError({ message }: { message?: string }) {
+    if (!message) return null;
+    return <p className="text-red-400 text-sm">{message}</p>;
+}
