+++ src/components/PageOrder/orderFormClasses.ts
@@ -1,0 +1,3 @@
+export const labelCls = "text-[#121212] font-inter font-semibold leading-normal uppercase fluid-text [--fmin:22] [--fmax:26]";
+
+export const formBlockCls = "flex flex-col items-stretch justify-stretch gap-5 w-full [&_label]:mb-2.5 [&_input]:grow [&_input]:w-full [&_input]:p-5 [&_input]:bg-[#F4F4F4] [&_input]:text-black [&_input]:font-inter [&_input]:font-normal [&_input]:leading-normal [&_input::placeholder]:text-[#898989] [&_input::placeholder]:font-inter [&_input::placeholder]:font-normal [&_input::placeholder]:leading-normal [&_input[type=radio]]:grow-0 [&_input[type=radio]]:w-5 [&_textarea]:grow [&_textarea]:w-full [&_textarea]:h-[200px] [&_textarea]:resize-none [&_textarea]:p-5 [&_textarea]:bg-[#F4F4F4] [&_textarea]:text-black [&_textarea]:font-inter [&_textarea]:font-normal [&_textarea]:leading-normal [&_textarea::placeholder]:text-[#898989] [&_textarea::placeholder]:font-inter [&_textarea::placeholder]:font-normal [&_textarea::placeholder]:leading-normal";
