+++ src/data/Other/OrderOptions.ts
@@ -1,0 +1,13 @@
+export const deliveryOptions = [
+    { value: "courier", label: "Courier" },
+    { value: "nova-poshta", label: "Nova Poshta" },
+    { value: "ukrposhta", label: "Ukrposhta" },
+    { value: "self-pickup", label: "Self-pickup" },
+];
+
+export const paymentOptions = [
+    { value: "Privat24", label: "Privat 24" },
+    { value: "Monobank", label: "Monobank" },
+    { value: "Abank", label: "Abank" },
+    { value: "Sense Bank", label: "Sense Bank" },
+];
